==> notebooks/0-peterdesmet-get-name-counts.ipynb <==
# Count different name types

import pandas as pd

data = pd.read_table('../data/interim/verified-checklist.tsv', dtype=object)

data.head()

## Number of records

data['index'].count()

## Number of *unique* usageKeys

Unmatched names will not be included.

data.groupby('gbifapi_usageKey').first().reset_index()['index'].count()

## Number of records that need synonymy verification

Might contain duplicates.

data[data['nameMatchValidation'].str.contains('verify', na=False)]['index'].count()

## Number of verified *unique* acceptedKeys

Only names that contain `ok`, so no wrong matches, no matches or unverified synonyms.

data[data['nameMatchValidation'].str.contains('ok', na=False)].groupby('gbifapi_usageKey').first().reset_index()['index'].count()
