==> notebooks/1-peterdesmet-create-aggregated-checklist.ipynb <==
# Created aggregated checklist

import pandas as pd
from collections import OrderedDict

data = pd.read_table('../data/interim/verified-checklist.tsv', dtype=object)

data.head()

## Retrieve valid records only

Only records that contain ok, so no wrong matches, no matches or unverified synonyms.

valid_records = data[data['nameMatchValidation'].str.contains('^ok', regex=True, na=False)]

valid_records['index'].count()

## Aggregate and sort on gbifapi_acceptedName & collect unique values for some columns

def split_delimited_values(series, delimiter='|'):
    # Splits values (of a series) containing a delimiter in multiple values
    # Is a bit of a hack
    
    # Only split if the series contains more than NaN values, otherwise, just return series back
    if len(series.dropna()) > 0:
        series = series.str.lower().str.split(delimiter).apply(pd.Series, 1).stack()
        series = series.str.strip()
        series.index = series.index.droplevel(-1)
    return series

def get_unique_values(series):
    # dropna(): remove NaN values
    # tuple(): will collect all values of a series as a tuple
    # set(): will only keep (unordered) unique values
    # Returns a set
    
    # It would have been easier to just return a list(), so we can sort it too,
    # but then we get 'Function does not reduce' (http://stackoverflow.com/a/37955931)
    # so have to use tuple()
    return set(tuple(series.dropna()))

# Default dict syntax for aggregation does not preserve column order, which is why we use OrderedDict
unique_values_per_column = OrderedDict([
    ('kingdom', lambda x: get_unique_values(x)),
    ('datasetName', lambda x: get_unique_values(x)),
    ('euConcernStatus', lambda x: get_unique_values(x)),
    ('firstObservationYearBE', lambda x: get_unique_values(x)),
    ('firstObservationYearFL', lambda x: get_unique_values(x)),
    ('invasionStage', lambda x: get_unique_values(x)),
    ('habitat', lambda x: get_unique_values(x)),
    ('nativeRange', lambda x: get_unique_values(x)),
    ('introductionPathway', lambda x: get_unique_values(split_delimited_values(x))), # Can contain |
    ('presenceBE', lambda x: get_unique_values(x)),
    ('presenceFL', lambda x: get_unique_values(x)),
    ('presenceWA', lambda x: get_unique_values(x)),
    ('presenceBR', lambda x: get_unique_values(x)),
    ('gbifapi_scientificName', lambda x: get_unique_values(x)),
    ('index', lambda x: get_unique_values(x))
])

aggregated_records = valid_records.groupby(['gbifapi_acceptedScientificName','gbifapi_acceptedKey']).agg(unique_values_per_column).reset_index()

aggregated_records = aggregated_records.sort_values(by='gbifapi_acceptedScientificName')

aggregated_records.head()

## Sort and concatenate unique values per column

def sort_and_concatenate(aSet, sortAs=str):
    # sortAs allows to sort more specific
    # {9, 200, 12} will be sorted as:
    # '12, 200, 9' with default str
    # '9, 12, 200' with int
    sortedList = sorted(aSet, key=sortAs)
    concatenatedList = ' | '.join(str(i) for i in sortedList)
    return concatenatedList # a string

aggregated_records['kingdom'] = aggregated_records['kingdom'].apply(lambda x: sort_and_concatenate(x))
aggregated_records['datasetName'] = aggregated_records['datasetName'].apply(lambda x: sort_and_concatenate(x))
aggregated_records['euConcernStatus'] = aggregated_records['euConcernStatus'].apply(lambda x: sort_and_concatenate(x))
aggregated_records['firstObservationYearBE'] = aggregated_records['firstObservationYearBE'].apply(lambda x: sort_and_concatenate(x))
aggregated_records['firstObservationYearFL'] = aggregated_records['firstObservationYearFL'].apply(lambda x: sort_and_concatenate(x))
aggregated_records['invasionStage'] = aggregated_records['invasionStage'].apply(lambda x: sort_and_concatenate(x))
aggregated_records['habitat'] = aggregated_records['habitat'].apply(lambda x: sort_and_concatenate(x))
aggregated_records['nativeRange'] = aggregated_records['nativeRange'].apply(lambda x: sort_and_concatenate(x))
aggregated_records['introductionPathway'] = aggregated_records['introductionPathway'].apply(lambda x: sort_and_concatenate(x))
aggregated_records['presenceBE'] = aggregated_records['presenceBE'].apply(lambda x: sort_and_concatenate(x))
aggregated_records['presenceFL'] = aggregated_records['presenceFL'].apply(lambda x: sort_and_concatenate(x))
aggregated_records['presenceWA'] = aggregated_records['presenceWA'].apply(lambda x: sort_and_concatenate(x))
aggregated_records['presenceBR'] = aggregated_records['presenceBR'].apply(lambda x: sort_and_concatenate(x))
aggregated_records['gbifapi_scientificName'] = aggregated_records['gbifapi_scientificName'].apply(lambda x: sort_and_concatenate(x))
aggregated_records['index'] = aggregated_records['index'].apply(lambda x: sort_and_concatenate(x, int))

aggregated_records.head()

## Write aggregated records to file

aggregated_records.to_csv('../data/processed/aggregated-checklist.tsv', sep='\t', index=False)
