==> notebooks/2-peterdesmet-merge-aggregated-checklist-with-shortlist.ipynb <==
import pandas as pd

agg_checklist = pd.read_table('../data/processed/aggregated-checklist.tsv', dtype=object)

shortlist = pd.read_table('../../invasive-t0-occurrences/data/vocabularies/species-list.tsv', dtype=object)

## Filter aggregated checklist on validated names

Not necessary, aggregated checklist only contains validated names

## Filter shortlist on validated names

valid_shortlist = shortlist[shortlist['nameMatchValidation'].str.contains('^ok', regex=True)]

## Merge shortlist with aggregated checklist

merged_list = pd.merge(valid_shortlist, agg_checklist, how='left', 
         left_on='gbifapi_acceptedKey', right_on='gbifapi_acceptedKey', 
         left_index=False, right_index=False, sort=True, 
         suffixes=('_shortlist', '_agg'), copy=True, indicator=False)

merged_list.head()

## Missing names in aggregated checklist

missing_names = merged_list[merged_list['index'].isnull()]

len(missing_names)

missing_names[['name', 'euConcernStatus']]

Remark by Tim:

> Looks OK, most of the species above do not occur in Belgium, except for `Vespa velutina`.
